Normalize error and add history hooks
The error report can now be an array of strings even for array and
object values. In this case, the errors are the ones that apply to the
entire value, similar to the ones stored at the `""` key.

New convenience functions help extract the global errors and the
children errors from an `error` object: `globalError`, `itemError`, and
`propertyError`.

New hooks enable getting the previous value of a property, or the
previous argument values passed on to a function: `usePrevious`,
`usePreviousArgument`, and `usePreviousArgumentList`.